+++ src/config/add-dependent-types-to-only-types.ts
@@ -11,16 +11,36 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-import { TypeDefinitionNode } from "graphql/index";
+import { CodegenConfigWithDefaults } from "./build-config-with-defaults";
 import { GraphQLSchema } from "graphql";
+import { TypeDefinitionNode } from "graphql/index";
 import {
   getDependentFieldTypeNames,
   getDependentInterfaceNames,
   getDependentUnionNames,
-} from "./dependent-type-utils";
-import { CodegenConfigWithDefaults } from "./build-config-with-defaults";
+} from "../utils/dependent-type-utils";
 
-export function getDependentTypeNames(
+export function addDependentTypesToOnlyTypes(
+  config: CodegenConfigWithDefaults,
+  schema: GraphQLSchema,
+) {
+  if (!config.onlyTypes) {
+    throw new Error(`onlyTypes config is required to add dependent types`);
+  }
+  const onlyTypesNodes = config.onlyTypes
+    .map((typeName) => schema.getType(typeName)?.astNode)
+    .filter(Boolean);
+  const dependentTypeNames = onlyTypesNodes.flatMap((node) =>
+    getDependentTypeNames(schema, node, config),
+  );
+  const typesInScope = config.dependentTypesInScope;
+  const dependentTypesInScope = typesInScope
+    ? dependentTypeNames.filter((typeName) => typesInScope.includes(typeName))
+    : dependentTypeNames;
+  config.onlyTypes.push(...dependentTypesInScope);
+}
+
+function getDependentTypeNames(
   schema: GraphQLSchema,
   node: TypeDefinitionNode,
   config: CodegenConfigWithDefaults,
