+++ src/definitions/interface.ts
@@ -12,12 +12,12 @@ limitations under the License.
 */
 
 import { GraphQLSchema, InterfaceTypeDefinitionNode } from "graphql";
-import { buildAnnotations } from "../helpers/build-annotations";
-import { buildTypeMetadata } from "../helpers/build-type-metadata";
-import { shouldExcludeTypeDefinition } from "../helpers/should-exclude-type-definition";
-import { buildFieldDefinition } from "../helpers/build-field-definition";
-import { CodegenConfigWithDefaults } from "../helpers/build-config-with-defaults";
-import { getDependentInterfaceNames } from "../helpers/dependent-type-utils";
+import { buildAnnotations } from "../annotations/build-annotations";
+import { buildTypeMetadata } from "../utils/build-type-metadata";
+import { shouldExcludeTypeDefinition } from "../config/should-exclude-type-definition";
+import { buildFieldDefinition } from "./field";
+import { CodegenConfigWithDefaults } from "../config/build-config-with-defaults";
+import { getDependentInterfaceNames } from "../utils/dependent-type-utils";
 
 export function buildInterfaceDefinition(
   node: InterfaceTypeDefinitionNode,
