+++ src/utils/input-type-has-matching-output-type.ts
